Create: Finding Roots in Quadratic equation
Creating a Java program that prints all real solutions to the quadratic equation a*x²+b*x+c = 0. Read in a, b, c and use the quadratic formula. 
If the discriminant b*b -4ac is negative, display a message stating that there are no real solutions.
If b*b - 4 ac is positive, print real roots if they are equal print roots are equal.